+++ modifiers/prismjs.test.js
@@ -16,7 +16,7 @@ limitations under the License.
 
 'use strict';
 
-const md = require('markdown-it')().use(require('./prismjs'));
+const md = require('markdown-it')().use(require('./prismjs'), { logging: true });
 
 it('handles a code fence with a language (adding the class to the pre + highlighting)', () => {
     expect(md.render('```nginx\nserver {\n    try_files test =404;\n}\n```')).toBe(`<pre class="language-nginx"><code class="language-nginx"><span class="token directive"><span class="token keyword">server</span></span> <span class="token punctuation">{</span>
@@ -33,11 +33,20 @@ it('handles a code fence with a language alias', () => {
 });
 
 it('does not repeatedly load a modifier component', () => {
-    const error = jest.spyOn(global.console, 'error');
-    md.render('```ts\nconsole.log("test");\n```\n\n```ts\nconsole.log("test");\n```');
-    expect(error).not.toHaveBeenCalledWith(expect.stringContaining('Failed to load Prism component'), 'js-templates', expect.anything());
-    error.mockReset();
-    error.mockRestore();
+    const callCount = jest.fn();
+    jest.doMock('../vendor/prismjs/components/prism-js-templates', () => {
+        const actual = jest.requireActual('../vendor/prismjs/components/prism-js-templates');
+        return (...args) => {
+            callCount();
+            return actual(...args);
+        };
+    });
+
+    const mdTemp = require('markdown-it')().use(require('./prismjs'), { logging: true });
+    mdTemp.render('```ts\nconsole.log("test");\n```\n\n```ts\nconsole.log("test");\n```');
+    expect(callCount).toHaveBeenCalledTimes(1);
+
+    jest.resetModules();
 });
 
 it('does not pollute global scope', () => {
@@ -49,6 +58,34 @@ it('does not pollute global scope', () => {
 
     expect(window.Prism).toBeUndefined();
     expect(global.Prism).toBeUndefined();
+});
+
+describe('Error logging', () => {
+    const mdNoLogging = require('markdown-it')().use(require('./prismjs'), { logging: false });
+
+    it('does log errors when logging is enabled', () => {
+        const error = jest.spyOn(global.console, 'error');
+        jest.doMock('../vendor/prismjs/components/prism-typescript', () => { throw new Error('test'); });
+
+        md.render('```ts\nconsole.log("test");\n```');
+        expect(error).toHaveBeenCalledWith('Failed to load Prism component', 'typescript', expect.any(Error));
+
+        error.mockReset();
+        error.mockRestore();
+        jest.resetModules();
+    });
+
+    it('does not log errors when logging is disabled', () => {
+        const error = jest.spyOn(global.console, 'error');
+        jest.doMock('../vendor/prismjs/components/prism-typescript', () => { throw new Error('test'); });
+
+        mdNoLogging.render('```ts\nconsole.log("test");\n```');
+        expect(error).not.toHaveBeenCalled();
+
+        error.mockReset();
+        error.mockRestore();
+        jest.resetModules();
+    });
 });
 
 describe('HTML preservation', () => {